We now also download package artifacts
