+++ tools/utils/__init__.py
@@ -11,6 +11,7 @@ import sys
 import tempfile
 import zipfile
 from datetime import datetime
+from enum import IntEnum
 from typing import Any
 
 import packaging.version
@@ -31,6 +32,12 @@ SPB_ENVIRONMENT = os.environ.get("SPB_ENVIRONMENT") or "test"
 STAGING_BUCKET_NAME = f"salt-project-{SPB_ENVIRONMENT}-salt-artifacts-staging"
 RELEASE_BUCKET_NAME = f"salt-project-{SPB_ENVIRONMENT}-salt-artifacts-release"
 BACKUP_BUCKET_NAME = f"salt-project-{SPB_ENVIRONMENT}-salt-artifacts-backup"
+
+
+class ExitCode(IntEnum):
+    OK = 0
+    FAIL = 1
+    SOFT_FAIL = 2
 
 
 def create_progress_bar(file_progress: bool = False, **kwargs):
@@ -200,84 +207,7 @@ def get_file_checksum(fpath: pathlib.Path, hash_name: str) -> str:
     return hexdigest
 
 
-def get_github_token(ctx: Context) -> str | None:
-    """
-    Get the GITHUB_TOKEN to be able to authenticate to the API.
-    """
-    github_token = os.environ.get("GITHUB_TOKEN")
-    if github_token is None:
-        gh = shutil.which("gh")
-        ret = ctx.run(gh, "auth", "token", check=False, capture=True)
-        if ret.returncode == 0:
-            github_token = ret.stdout.decode().strip() or None
-    return github_token
-
-
-def download_artifact(
-    ctx: Context,
-    dest: pathlib.Path,
-    run_id: int,
-    repository: str = "saltstack/salt",
-    artifact_name: str | None = None,
-):
-    """
-    Download CI artifacts.
-    """
-    github_token = get_github_token(ctx)
-    if github_token is None:
-        ctx.error("Downloading artifacts requires being authenticated to GitHub.")
-        ctx.info(
-            "Either set 'GITHUB_TOKEN' to a valid token, or configure the 'gh' tool such that "
-            "'gh auth token' returns a token."
-        )
-        ctx.exit(1)
-    with ctx.web as web:
-        headers = {
-            "Accept": "application/vnd.github+json",
-            "Authorization": f"Bearer {github_token}",
-        }
-        web.headers.update(headers)
-        page = 0
-        found_artifact = False
-        while True:
-            if found_artifact:
-                break
-            page += 1
-            ret = web.get(
-                f"https://api.github.com/repos/{repository}/actions/runs/{run_id}/artifacts?per_page=100&page={page}"
-            )
-            if ret.status_code != 200:
-                ctx.error(
-                    f"Failed to get the artifacts for the run ID {run_id} for repository {repository!r}: {ret.reason}"
-                )
-                ctx.exit(1)
-            data = ret.json()
-            if not data["artifacts"]:
-                break
-            for artifact in data["artifacts"]:
-                if fnmatch.fnmatch(artifact["name"], artifact_name):
-                    found_artifact = artifact["name"]
-                    tempdir_path = pathlib.Path(tempfile.gettempdir())
-                    download_url = artifact["archive_download_url"]
-                    ctx.info(f"Downloading {download_url}")
-                    downloaded_artifact = _download_file(
-                        ctx,
-                        download_url,
-                        tempdir_path / f"{artifact['name']}.zip",
-                        headers=headers,
-                    )
-                    ctx.info("Downloaded", downloaded_artifact)
-                    with zipfile.ZipFile(downloaded_artifact) as zfile:
-                        zfile.extractall(path=dest)
-                    break
-        if found_artifact is False:
-            ctx.error(f"Failed to find an artifact by the name of {artifact_name!r}")
-            ctx.exit(1)
-
-    return found_artifact
-
-
-def _download_file(
+def download_file(
     ctx: Context,
     url: str,
     dest: pathlib.Path,
